lib/python/fun/scrop_cr2_to_cr4.py BUGFIX: float instead of int is used for fileparts.

+++ trunk/lib/python/fun/scrop_cr2_to_cr4.py
@@ -4,7 +4,7 @@ filesize=os.path.getsize(sobj.crop.Data_output_file);
 width=sobj.crop.Last_pixel-sobj.crop.First_pixel+1;
 datatype, complexFlag=adore.dataFormat2dataType('cr2');
 length=float(filesize)/width/np.dtype(datatype).itemsize/2; 
-fileparts=fix(linspace(0, length, 10));
+fileparts=fix(linspace(0, length, 10)).astype(int);
   
 #adore.writedata(sobj.crop.Data_output_file+'.cr4', data, 'cr4')
 fout=open(sobj.crop.Data_output_file+'.cr4', 'wb');
